修复webp构建未通过问题

+++ frameworks/innerkitsimpl/accessor/src/webp_exif_metadata_accessor.cpp
@@ -481,7 +481,7 @@ bool WebpExifMetadataAccessor::WriteHeader(BufferMetadataStream &bufStream, uint
     fileSize -= exifData.Size() % WEBP_BUF_SIZE ? (exifData.Size() + 1) : exifData.Size();
     UL2Data(headInfo.Data(WEBP_FILE_SIZE_BUFF_SIZE), fileSize, littleEndian);
     IMAGE_LOGD("Write webp file size: %{public}u, old exif size: %{public}u, new exif size: %{public}lu",
-        fileSize, size, exifData.Size());
+        fileSize, size, static_cast<unsigned long>(exifData.Size()));
     return bufStream.Write(headInfo.Data(), headInfo.Size()) == (ssize_t)headInfo.Size();
 }
